OrderManager/src/requirements.txt
fastapi
uvicorn
pydantic
tinkoff-investments
